renamed test to app
